pinmux_helper_imx6: refactor project structure
move code to src directory, which is a preperational step for deployment
and improves import handling for pytest.

add separate callable main function, which can be used as entrypoint in
the pyproject.toml added in a later commit

Fixup the import paths in the pytest tests.
